Module Conduit_tls.Make

Parameters

IO : Conduit.IO

Conduit : Conduit.S with type input = Cstruct.t and type output = Cstruct.t and type +'a io = 'a IO.t

Signature

type 'flow protocol_with_tls

val underlying : 'flow protocol_with_tls -> 'flow

underlying flow returns underlying flow used by the TLS flow.

val handshake : 'flow protocol_with_tls -> bool

handshake flow returns true if handshake is processing.

val protocol_with_tls : ('edn, 'flow) Conduit.protocol -> ('edn * Tls.Config.client, 'flow protocol_with_tls) Conduit.protocol

From a given protocol witness, it creates a new witness of the protocol layered with TLS.

type 'service service_with_tls

val service_with_tls : ('cfg, 't, 'flow) Conduit.Service.service -> ('edn, 'flow protocol_with_tls) Conduit.protocol -> ('cfg * Tls.Config.server, 't service_with_tls, 'flow protocol_with_tls) Conduit.Service.service