Module type 2-Conduit.FLOW

A flow is a system that allows entities to transmit payloads. These entities do not have to care about the underlying transport mechanism. flows simply deal with routing and delivering of these payloads. That abstraction allows these protocols to compose.

For example, the Transmission Control Protocol (TCP) is representable as a flow, because it is able to encapsulate some payloads without interpreting it. A counter-example is the Simple Mail Transfer Protocol (SMTP) which needs an interpretation of its payloads: tokens such as EHLO or QUIT have a direct incidence over the life-cycle of the connection.

An other protocol representable as a flow is the Transport Layer Security (TLS), as it deals only with privacy and data integrity. Conduit is able to compose flows together like TCP ∘ TLS to make a new flow. Higher-level protocols can be built in top of these abstract flows: for instance, Secure Simple Mail Transfer Protocol (SSMTP) or HyperText Transfer Protocol Secure (HTTPS) can be defined on top of both TCP and TLS. Using Conduit, these can be abstracted to work over any flow implementations.

type +'a io = 'a io

type flow

Input & Output.

Depending on the I/O model, the type for inputs and outputs can differ ; for instance they could allow users the ability to define capabilities on them such as read or write capabilities.

However, in most of the current Conduit backends:

type input = Cstruct.t
type output = Cstruct.t

type input = input

and output = output

Errors.

type error

The type for errors.

val pp_error : error Fmt.t

pp_error is the pretty-printer for error.

val recv : flow -> input -> (int Conduit__.Sigs.or_end_of_flow, error) Stdlib.result io

recv flow input is Ok (`Input len) iff len bytes of data has been received from the flow flow and copied in input.

val send : flow -> output -> (int, error) Stdlib.result io

send t output is Ok len iff len bytes of data from output has been sent over the flow flow.

val close : flow -> (unit, error) Stdlib.result io

close flow closes flow. Subsequent calls to recv on flow will return `End_of_flow. Subsequent calls to send on t will return an Error.